+++ epsic_tools/api.py
@@ -11,4 +11,4 @@ from epsic_tools.toolbox import radial_profile
 from epsic_tools.toolbox import warp_3d
 from epsic_tools.toolbox import make_merlin_mask
 from epsic_tools.toolbox import sim_utils
-from epsic_tools.toolbox import recon_utils
+from epsic_tools.toolbox import ptycho_utils
